Turns out go is not C or ruby
So slices don't just work as easy as in ruby
But i have no idea what is happening in detail either.

Seems like they are structs and somehow a mess.

Using fixed arrays instead and a bunch of compiler driven try and error
I fixed the bug of the server token getting corrupted.